# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/settings": {
      "get": {
        "operationId": "SettingsController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "تنظیمات سامانه + بازه‌های واقعی جریمهٔ استرداد (فاز ۷)",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "به‌روزرسانی تنظیمات — کلیدهای ناشناخته رد می‌شوند",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/support-tickets": {
      "post": {
        "operationId": "SupportTicketsController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "ثبت تیکت پشتیبانی (بدون ورود به حساب)",
        "tags": [
          "support-tickets"
        ]
      },
      "get": {
        "operationId": "SupportTicketsController_list",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dept",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فهرست تیکت‌های پشتیبانی",
        "tags": [
          "support-tickets"
        ]
      }
    },
...
    "/support-tickets/{id}": {
      "get": {
        "operationId": "SupportTicketsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "جزئیات تیکت پشتیبانی",
        "tags": [
          "support-tickets"
        ]
      }
    },
    "/support-tickets/{id}/forward": {
      "patch": {
        "operationId": "SupportTicketsController_forward",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForwardTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ارجاع تیکت به کارمند/مدیر",
        "tags": [
          "support-tickets"
        ]
      }
    },
    "/support-tickets/{id}/status": {
      "patch": {
        "operationId": "SupportTicketsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "تغییر وضعیت تیکت",
        "tags": [
          "support-tickets"
        ]
      }
    },
...
    "/my/support-tickets/{id}": {
      "get": {
        "operationId": "MySupportTicketsController_getMine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "جزئیات یک تیکت پشتیبانی مشتری",
        "tags": [
          "support-tickets"
        ]
      }
    },
...
    "/survey/settings": {
      "get": {
        "operationId": "SurveyController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "وضعیت و عنوان نظرسنجی",
        "tags": [
          "survey"
        ]
      },
      "patch": {
        "operationId": "SurveyController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSurveySettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فعال/غیرفعال یا تغییر عنوان نظرسنجی",
        "tags": [
          "survey"
        ]
      }
    },
    "/survey/questions": {
      "get": {
        "operationId": "SurveyController_listQuestions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فهرست سؤالات نظرسنجی",
        "tags": [
          "survey"
        ]
      },
      "post": {
        "operationId": "SurveyController_addQuestion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSurveyQuestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "افزودن سؤال نظرسنجی",
        "tags": [
          "survey"
        ]
      }
    },
    "/survey/questions/{id}": {
      "delete": {
        "operationId": "SurveyController_removeQuestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "حذف سؤال نظرسنجی",
        "tags": [
          "survey"
        ]
      }
    },
...
    "/survey/results/{flightInstanceId}/analyze": {
      "post": {
        "operationId": "SurveyController_analyze",
        "parameters": [
          {
            "name": "flightInstanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "خلاصهٔ هوش مصنوعی نظرات یک پرواز",
        "tags": [
          "survey"
        ]
      }
    },
...
  },
  "info": {
    "title": "blujet API",
    "description": "blujet airline platform — internal API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateSettingsDto": {
        "type": "object",
        "properties": {
          "patch": {
            "type": "object",
            "description": "Partial key-value patch — unknown keys are rejected",
            "example": {
              "maintenance": true
            }
          }
        },
        "required": [
          "patch"
        ]
      },
...
      "SubmitSupportTicketDto": {
        "type": "object",
        "properties": {
          "requesterName": {
            "type": "string",
            "example": "نگار رضایی",
            "description": "نام و نام خانوادگی"
          },
          "requesterPhone": {
            "type": "string",
            "example": "09121234567",
            "description": "شماره تماس"
          },
          "subject": {
            "type": "string",
            "example": "مشکل در پرداخت",
            "description": "موضوع تیکت"
          },
          "body": {
            "type": "string",
            "example": "توضیح مشکل...",
            "description": "متن تیکت"
          }
        },
        "required": [
          "requesterName",
          "requesterPhone",
          "subject",
          "body"
        ]
      },
      "ForwardTicketDto": {
        "type": "object",
        "properties": {
          "targetUserId": {
            "type": "string",
            "description": "شناسه کارمند/مدیر مقصد ارجاع"
          }
        },
        "required": [
          "targetUserId"
        ]
      },
      "UpdateTicketStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "ANSWERED",
              "CLOSED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdateSurveySettingsDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "فعال/غیرفعال بودن نظرسنجی"
          },
          "title": {
            "type": "string",
            "example": "نظرسنجی رضایت مسافران"
          }
        }
      },
      "CreateSurveyQuestionDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "راحتی صندلی و کابین"
          }
        },
        "required": [
          "label"
        ]
      },
...
    }
  }
}